+++ man-roxygen/params_prim_lhs_rhs_int.R
@@ -1,4 +1,0 @@
-#' @param lhs,rhs ([`arrayish`])\cr
-#'   Arrayish values of data type integer or unsigned integer. A boolean is not
-#'   accepted: the StableHLO spec's *tensor of integer type* excludes `i1`.
-#'   Must have the same shape.